+++ NEWS.md
@@ -11,6 +11,7 @@
 * Data import log not sorted issues ([MODSOURMAN-1433](https://folio-org.atlassian.net/browse/MODSOURMAN-1433))
 * Import of MARC bib with 999 ff $s (only) completes without error ([MODSOURMAN-1447](https://folio-org.atlassian.net/browse/MODSOURMAN-1447))
 * Fix "invalid input syntax for type integer" error in get_job_log_entries for invoices with alphanumeric vendor invoice numbers ([MODSOURMAN-1346](https://folio-org.atlassian.net/browse/MODSOURMAN-1346))
+* Fix log entries sorting by items/holdings ([MODSOURMAN-1453](https://folio-org.atlassian.net/browse/MODSOURMAN-1453))
 
 ### Tech Dept